WIP: Move button controls
